fix: Generalize the tracing provider
