+++ pyworkflow/tracing/__init__.py
@@ -4,11 +4,12 @@ Tracing module for PyWorkflow.
 Provides observability tracing (e.g. Langfuse) for workflows and steps.
 """
 
-from pyworkflow.tracing.provider import TracingProvider, create_tracing_provider
+from pyworkflow.tracing.base import BaseTracingProvider
+from pyworkflow.tracing.factory import create_tracing_provider
 from pyworkflow.tracing.types import LLMCallData, StepTracingData, ToolCallData
 
 __all__ = [
-    "TracingProvider",
+    "BaseTracingProvider",
     "create_tracing_provider",
     "LLMCallData",
     "StepTracingData",
